refactor: move legacy profile config types into tokmd-settings
